added convertNames function
