+++ api/get.js
@@ -5,6 +5,7 @@ var router = express.Router();
 // functions called by API endpoints
 const requisiteHelpers = require('../helpers/checkRequisites.js');
 const dbHelpers = require('../helpers/db.js');
+const helperFunctions = require('../helpers/functions');
 
 router.get('/test/:name', (req,res  )=> {
   res.send("Welcome, " + req.params.name);
@@ -64,7 +65,7 @@ router.get('/validMajorClasses', async function(req,res,next){
 
 
 	var major = req.body.department;
-	var takenCourses = req.body.takenCourses;
+	var takenCourses = helperFunctions.convertNames(req.body.takenCourses);
 	try{
 		var objectArray = await courses.find({"department":major}).toArray();
 	} catch(error){
@@ -149,4 +150,95 @@ router.get('/checkRequisites', function(req,res,next){
 		})
 	return;
 });
+*/
+
+
+
+/*
+
+REQUEST BODY for validMajorClasses
+
+{  
+   "department":"Computer Science",
+   "takenCourses":{  
+      "classes":{  
+         "c1":{  
+            "id":"c1",
+            "dept":"COM SCI",
+            "name":"1"
+         },
+         "c2":{  
+            "id":"c2",
+            "dept":"COM SCI",
+            "name":"31"
+         },
+         "c3":{  
+            "id":"c3",
+            "dept":"ENGCOMP",
+            "name":"3"
+         },
+         "c4":{  
+            "id":"c4",
+            "dept":"MATH",
+            "name":"31A"
+         },
+         "c5":{  
+            "id":"c5",
+            "dept":"PHYSICS",
+            "name":"1B"
+         },
+         "c6":{  
+            "id":"c6",
+            "dept":"PHYSICS",
+            "name":"1C"
+         }
+      },
+      "quarters":{  
+         "search":{  
+            "id":"search",
+            "title":"To Do",
+            "classIds":[  
+               "c5",
+               "c6"
+            ]
+         },
+         "q1":{  
+            "id":"q1",
+            "title":"To Do",
+            "classIds":[  
+               "c1",
+               "c2"
+            ]
+         },
+         "q2":{  
+            "id":"q2",
+            "title":"To Do",
+            "classIds":[  
+               "c4"
+            ]
+         },
+         "q3":{  
+            "id":"q3",
+            "title":"To Do",
+            "classIds":[  
+
+            ]
+         },
+         "q4":{  
+            "id":"q4",
+            "title":"To Do",
+            "classIds":[  
+               "c3"
+            ]
+         }
+      },
+      "quarterOrder":[  
+         "q1",
+         "q2",
+         "q3",
+         "q4"
+      ]
+   }
+}
+
 */
