+++ utils/acronym_mapping.json
@@ -1,0 +1,193 @@
+{
+    "AERO ST": "Aerospace Studies",
+    "ARCHEOL": "Archaeology",
+    "AF AMER": "African American Studies",
+    "ARCH&UD": "Architecture and Urban Design",
+    "AFRC ST": "African Studies",
+    "ARMENIA": "Armenian",
+    "AFRKAAN": "Afrikaans",
+    "ART": "Art",
+    "AM IND": "American Indian Studies",
+    "ART HIS": "Art History",
+    "ASL": "American Sign Language",
+    "ART&ARC": "Arts and Architecture",
+    "AN N EA": "Ancient Near East",
+    "ARTS ED": "Arts Education",
+    "ANES": "Anesthesiology",
+    "ASIAN": "Asian",
+    "ANTHRO": "Anthropology",
+    "ASIA AM": "Asian American Studies",
+    "APPLING": "Applied Linguistics",
+    "ASTR": "Astronomy",
+    "ARABIC": "Arabic",
+    "A&O SCI": "Atmospheric and Oceanic Sciences",
+    "BIOENGR": "Bioengineering",
+    "BIOMATH": "Biomathematics",
+    "BIOINFO": "Bioinformatics (Graduate)",
+    "BMD RES": "Biomedical Research",
+    "BIOINFR": "Bioinformatics (Undergraduate)",
+    "BIOSTAT": "Biostatistics",
+    "BIOL CH": "Biological Chemistry",
+    "BULGR": "Bulgarian",
+    "C&EE ST": "Central and East European Studies",
+    "CLUSTER": "Clusters",
+    "CH ENGR": "Chemical Engineering",
+    "COMM": "Communication",
+    "CHEM": "Chemistry and Biochemistry",
+    "COM HLT": "Community Health Sciences",
+    "CHICANO": "Chicana and Chicano Studies",
+    "COM LIT": "Comparative Literature",
+    "CHIN": "Chinese",
+    "C&S BIO": "Computational and Systems Biology",
+    "CIVIC": "Civic Engagement",
+    "COM SCI": "Computer Science",
+    "C&EE": "Civil and Environmental Engineering",
+    "CAEM": "Conservation of Archaeological and Ethnographic Materials",
+    "CLASSIC": "Classics",
+    "CZCH": "Czech",
+    "DANCE": "Dance",
+    "DGT HUM": "Digital Humanities",
+    "DENT": "Dentistry",
+    "DIS STD": "Disability Studies",
+    "DESMA": "Design / Media Arts",
+    "DUTCH": "Dutch",
+    "EPS SCI": "Earth, Planetary, and Space Sciences",
+    "ENGL": "English",
+    "EA STDS": "East Asian Studies",
+    "ESL": "English as A Second Language",
+    "EE BIOL": "Ecology and Evolutionary Biology",
+    "ENGCOMP": "English Composition",
+    "ECON": "Economics",
+    "ENVIRON": "Environment",
+    "EDUC": "Education",
+    "ENV HLT": "Environmental Health Sciences",
+    "EC ENGR": "Electrical and Computer Engineering",
+    "EPIDEM": "Epidemiology",
+    "ENGR": "Engineering",
+    "ETHNMUS": "Ethnomusicology",
+    "FAM MED": "Family Medicine",
+    "FOOD ST": "Food Studies",
+    "FILIPNO": "Filipino",
+    "FRNCH": "French",
+    "FILM TV": "Film and Television",
+    "GENDER": "Gender Studies",
+    "GJ STDS": "Global Jazz Studies",
+    "GEOG": "Geography",
+    "GLBL ST": "Global Studies",
+    "GERMAN": "German",
+    "GRAD PD": "Graduate Student Professional Development",
+    "GRNTLGY": "Gerontology",
+    "GREEK": "Greek",
+    "GLB HLT": "Global Health",
+    "HLT POL": "Health Policy and Management",
+    "HNRS": "Honors Collegium",
+    "HEBREW": "Hebrew",
+    "HUM GEN": "Human Genetics",
+    "HIN-URD": "Hindi-Urdu",
+    "HNGAR": "Hungarian",
+    "HIST": "History",
+    "IL AMER": "Indigenous Languages of the Americas",
+    "INTL DV": "International Development Studies",
+    "I E STD": "Indo-European Studies",
+    "I M STD": "International Migration Studies",
+    "INDO": "Indonesian",
+    "IRANIAN": "Iranian",
+    "INF STD": "Information Studies",
+    "ISLM ST": "Islamic Studies",
+    "I A STD": "International and Area Studies",
+    "ITALIAN": "Italian",
+    "JAPAN": "Japanese",
+    "JEWISH": "Jewish Studies",
+    "KOREA": "Korean",
+    "LBR&WS": "Labor and Workplace Studies",
+    "LGBTQS": "Lesbian, Gay, Bisexual, Transgender, and Queer Studies",
+    "LATIN": "Latin",
+    "LIFESCI": "Life Sciences",
+    "LATN AM": "Latin American Studies",
+    "LING": "Linguistics",
+    "LAW": "Law",
+    "LTHUAN": "Lithuanian",
+    "UG-LAW": "Law (Undergraduate)",
+    "MGMT": "Management",
+    "MED": "Medicine",
+    "MGMTEX": "Management-Executive MBA",
+    "MIMG": "Microbiology, Immunology, and Molecular Genetics",
+    "MGMTFT": "Management-Full-Time MBA",
+    "M E STD": "Middle EasternStudies",
+    "MGMTFE": "Management-Fully Employed MBA",
+    "MIL SCI": "Military Science",
+    "MGMTGEX": "Management-Global Executive MBA Asia Pacific",
+    "M PHARM": "Molecular and Medical Pharmacology",
+    "MGMTMFE": "Management-Master of Financial Engineering",
+    "MOL BIO": "Molecular Biology",
+    "MGMTMSA": "Management-Master of Science in Business Analytics",
+    "MOL TOX": "Molecular Toxicology",
+    "MGMTPHD": "Management-PhD",
+    "MCD BIO": "Molecular, Cell, and Developmental Biology",
+    "MAT SCI": "Materials Science and Engineering",
+    "MC&IP": "Molecular, Cellular, and Integrative Physiology",
+    "MATH": "Mathematics",
+    "MUSC": "Music",
+    "MECH&AE": "Mechanical and Aerospace Engineering",
+    "MSC IND": "Music Industry",
+    "MED HIS": "Medical History",
+    "MUSCLG": "Musicology",
+    "NAV SCI": "Naval Science",
+    "NEUROSC": "Neuroscience",
+    "NR EAST": "Near Eastern Languages",
+    "NEURO": "Neuroscience (Graduate)",
+    "NEURBIO": "Neurobiology",
+    "NEURSGY": "Neurosurgery",
+    "NEURLGY": "Neurology",
+    "NURSING": "Nursing",
+    "OBGYN": "Obstetrics and Gynecology",
+    "ORL BIO": "Oral Biology",
+    "OPTH": "Ophthalmology",
+    "ORTHPDC": "Orthopaedic Surgery",
+    "PATH": "Pathologyand Laboratory Medicine",
+    "POL SCI": "Political Science",
+    "PEDS": "Pediatrics",
+    "PORTGSE": "Portuguese",
+    "PHILOS": "Philosophy",
+    "COMPTNG": "Program in Computing",
+    "PHYSICS": "Physics",
+    "PSYCTRY": "Psychiatry and Biobehavioral Sciences",
+    "PBMED": "Physics and Biology in Medicine",
+    "PSYCH": "Psychology",
+    "PHYSCI": "Physiological Science",
+    "PUB AFF": "Public Affairs",
+    "PHYSIOL": "Physiology",
+    "PUB HLT": "Public Health",
+    "POLSH": "Polish",
+    "PUB PLC": "Public Policy",
+    "RAD ONC": "Radiation Oncology",
+    "ROMANIA": "Romanian",
+    "RELIGN": "Religion, Study of",
+    "RUSSN": "Russian",
+    "SCAND": "Scandinavian",
+    "SOC GEN": "Society and Genetics",
+    "SCI EDU": "Science Education",
+    "SOCIOL": "Sociology",
+    "SEMITIC": "Semitic",
+    "S ASIAN": "South Asian",
+    "SRB CRO": "Serbian/Croatian",
+    "SEASIAN": "Southeast Asian",
+    "SLAVC": "Slavic",
+    "SPAN": "Spanish",
+    "SOC SC": "Social Science",
+    "STATS": "Statistics",
+    "SOC THT": "Social Thought",
+    "SURGERY": "Surgery",
+    "SOC WLF": "SocialWelfare",
+    "SWAHILI": "Swahili",
+    "THAI": "Thai",
+    "TURKIC": "Turkic Languages",
+    "THEATER": "Theater",
+    "UKRN": "Ukrainian",
+    "URBN PL": "Urban Planning",
+    "UNIV ST": "University Studies",
+    "UROLOGY": "Urology",
+    "VIETMSE": "Vietnamese",
+    "WL ARTS": "World Arts and Cultures",
+    "YIDDSH": "Yiddish"
+}
